backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-multipart
pyjwt
passlib[bcrypt]
pyotp
faker

